Calculation updated

+++ Wearable/src/main/java/com/example/android/wearable/watchface/watchface/CalculationFatigueStress.java
@@ -75,9 +75,9 @@ public class CalculationFatigueStress {
         int Timeline = 300;
         int Tired = 0;
         int LastTired = 0;
-        int CurrentTired;
+        double CurrentTired;
 
-        CurrentTired = (int)slope(0, getMean(hrf), Timeline, getMean(hrl)*100);
+        CurrentTired = slope(0, getMean(hrf), Timeline, getMean(hrl)*100);
         if(CurrentTired > 0){
             if(CurrentTired > MaxTiredUp){
                 MaxTiredUp = (int)CurrentTired;
@@ -95,6 +95,7 @@ public class CalculationFatigueStress {
                 MinTiredDn = (int)CurrentTired;
             }
         }
+
         // 약간 이상한데?
         if(CurrentTired > 0){
             Tired = (int)((CurrentTired/(MaxTiredUp - MinTiredUp)) * 100);
